requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
pytrends==4.9.0
praw==7.7.1
requests==2.31.0
python-jose==3.3.0
apify-client==1.4.1
pandas==2.1.3 